Reworked option() configuration and initialization via app_init()
The use of the option() function provides a better more configurable way of working with global option settings.

This commit separates the Limonade dir [ option('limonade.dir.*') ] configurations from the App [ option('app.dir.*') ] configurations.

The general concept is to provide the possibility of different locations for the Limonade code base and that of the App specific codebase.

Also did the following changes:

* renamed option('session') to option('session.enabled') and option('session.name')
* renamed option('debug') to option('debug.enabled')