Added option('tmp.dir.views'..) support
It was too hard to change the option('dir.views') path when required. This minor change makes it much easier and does NOT mess with the option('dir.views') path.

The concept is simple. If you need to load a view from a different directory than the default directory, then it is a temporary change of path, so call it that way, and then handle its existence.