Added debug_settings() function (reworked settings-output)
Reworked the settings-output version and embedded it as a default functionality.

Why? Because one simple configuration and a simple function call in the bottom of your layout template provides a full overview of the variables and settings of the current app. Saves time!

Work-in-Progress:  should convert functionality to handle adding output values to the output dynamically.